added basics and started coding
I created two folders so far. One for the login site (begin) and one for the admin site.
Then i added the HTML tags for the version etc..
I started to build one of the two login containers already and gave it the headline as well as the input tags (Username and Password) and their labels.
I started at least with some of the background features for the body.